Grammar Study page, complete Exodus json, new Leviticus json
In AncientBiblicalHebrewGrammarStudy.php, identifying verb type and verbe tense is implemented. Same logic has been applied to weak verbs as strong verbs that this needs validation, studying and adjustment as I move along.

Exodus is completed with 40 chapters in youtube and info is stored in json file.

Leviticus has begun being loaded to YouTube and the info is being stored in json file.